pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-island"
version = "0.1.0"
description = "Prompt Island — Fully Autonomous AI Reality Show Backend"
requires-python = ">=3.11"
dependencies = [
    # ORM and database
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    # Data validation and structured outputs
    "pydantic>=2.9",
    # LLM providers
    "openai>=1.50",
    "anthropic>=0.40",
    # Environment configuration
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
